added isAuth to routes

+++ controllers/subjectsController.js
@@ -2,38 +2,44 @@
 const { db } = require("../models");
 const Subject = db.models.Subject;
 const SubjectUser = db.models.SubjectUser;
+const session = require("./sessionController");
 
 getAllSubjectsForUser = (req, res) => {
 	//get user from active user
-	const ownerId = 1; //for testing purposes
-	try {
-		Subject.findAll({
-			where: { ownerId },
-		}).then((subjects) => {
-			res.status(200).json(subjects);
-		});
-	} catch (err) {
-		res.send({ error: err });
-	}
-}; // works don't fuck with it
+	//const ownerId = 1; //for testing purposes
+	session.getUser().then(({id}) => {
+		try {
+			Subject.findAll({
+				where: { ownerId },
+			}).then((subjects) => {
+				res.status(200).json(subjects);
+			});
+		} catch (err) {
+			res.send({ error: err });
+		}
+	});
+};
 
 postNewSubject = (req, res) => {
-	const ownerId = 1; //get session user id
+	//const ownerId = 1; //get session user id
+	session.getUser().then(({id}) => {
 	const { title } = req.body;
-	try {
-		Subject.create({
-			title,
-			ownerId,
-		}).then(() => {
-			res.status(200).json({ message: `subject ${title} created` });
-		});
-	} catch (err) {
-		res.send({ error: err });
-	}
-}; //works don't fuck with it
+		try {
+			Subject.create({
+				title,
+				ownerId: id,
+			}).then(() => {
+				res.status(200).json({ message: `subject ${title} created` });
+			});
+		} catch (err) {
+			res.send({ error: err });
+		}
+	});
+};
 
 getSubject = (req, res) => {
 	const id = req.params.id; //subject_id
+	//if (Auth(req).auth) {} first get participants then add auth
 	try {
 		Subject.findAll({
 			where: { id },
@@ -46,7 +52,7 @@ getSubject = (req, res) => {
 	} catch (err) {
 		res.send({ error: err });
 	}
-}; //works. needs to return participants
+};
 
 putSubject = (req, res) => {
 	//advanced query
@@ -55,12 +61,14 @@ putSubject = (req, res) => {
 
 deleteSubject = (req, res) => {
 	const id = req.params.id;
-	try {
-		Subject.destroy({ where: { id } }).then(() => {
-			res.status(200).json({ message: `subject ${id} destroyed` });
-		});
-	} catch (err) {
-		res.send({ error });
+	if (Auth(req).auth) {
+		try {
+			Subject.destroy({ where: { id } }).then(() => {
+				res.status(200).json({ message: `subject ${id} destroyed` });
+			});
+		} catch (err) {
+			res.send({ error });
+		}
 	}
 }; // works. you know the drill...
